validate noop in secondary index `Update()` (#10806)

+++ go/libraries/doltcore/sqle/writer/prolly_index_writer.go
@@ -411,14 +411,27 @@ func (m prollySecondaryIndexWriter) Delete(ctx context.Context, sqlRow sql.Row) 
 	return m.mut.Delete(ctx, k)
 }
 
+func isNoopUpdate(oldRow, newRow sql.Row, keyMap val.OrdinalMapping) bool {
+	for to := range keyMap {
+		from := keyMap.MapOrdinal(to)
+		if oldRow[from] != newRow[from] {
+			return false
+		}
+	}
+	return true
+}
+
 func (m prollySecondaryIndexWriter) Update(ctx context.Context, oldRow sql.Row, newRow sql.Row) error {
+	// If no indexed columns are modified, no need to delete and update
+	if isNoopUpdate(oldRow, newRow, m.keyMap) {
+		return nil
+	}
+
 	oldKey, err := m.keyFromRow(ctx, oldRow)
 	if err != nil {
 		return err
 	}
 
-	// todo(andy): we can skip building, deleting |oldKey|
-	//  if we know the key fields are unchanged
 	if err := m.mut.Delete(ctx, oldKey); err != nil {
 		return err
 	}
